# specs/001-project-request-restaurant/contracts/api-contracts.yaml
openapi: 3.0.3
info:
  title: Restaurant Suite API
  description: Multi-role restaurant management and ordering system API
  version: 1.0.0
  contact:
    name: Restaurant Suite Team
servers:
  - url: https://api.restaurantsuite.example.com
    description: Production
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: Get orders
      description: Returns orders with optional filters (FR-032)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: tableId
          schema:
            type: string
            format: uuid
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Orders retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'

    post:
      tags: [Orders]
      summary: Create order
      description: Creates a new order (FR-023)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      description: Returns a specific order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Order retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/orders/{id}/status:
    put:
      tags: [Orders]
      summary: Update order status
      description: Transitions order through lifecycle (FR-026, FR-028)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
                notes:
                  type: string
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: Invalid status transition
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [orderLines]
      properties:
        tableId:
          type: string
          format: uuid
        notes:
          type: string
        orderLines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderLineRequest'
    OrderLineRequest:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        modifiers:
          type: array
          items:
            type: object
            properties:
              optionId:
                type: string
                format: uuid
              modifierId:
                type: string
                format: uuid
    OrderDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tableNumber:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        totalAmount:
          type: number
          format: decimal
        taxAmount:
          type: number
          format: decimal
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        orderLines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineDto'
    OrderLineDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        menuItemName:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: decimal
        modifiers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              priceDelta:
                type: number
                format: decimal
        status:
          type: string
          enum: [Pending, InProgress, Ready]
    OrderStatus:
      type: string
      enum: [Placed, Confirmed, InKitchen, Ready, Served, Completed, Cancelled]
    PaymentStatus:
      type: string
      enum: [Unpaid, PartiallyPaid, Paid, Refunded]

    # ===== Payments =====
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object